SetRoundingLLVM

Documentation for SetRoundingLLVM.

SetRoundingLLVM.llvm_rounding

SetRoundingLLVM.llvm_setrounding

SetRoundingLLVM.llvm_rounding — Method

llvm_get_rounding()

Return the current rounding mode for Float64

SetRoundingLLVM.llvm_setrounding — Method

llvm_set_rounding(round::RoundingMode)

Set the rounding mode of Float64, controlling the rounding of basic arithmetic functions (+, -, *, / and sqrt) and type conversion. Other numerical functions may give incorrect or invalid values when using rounding modes other than the default RoundNearest